Test fixes. Renamed multiple projects and fixed a test analysis problem for RT_Dispose.

+++ COM/COM2_test/Program.cs
@@ -33,7 +33,7 @@ namespace COM
 			int numBytes = serialPort.BytesToRead;
             serialPort.Read(m_recvBuffer, 0, numBytes);
 			for (int i = 0; i<numBytes; i++){
-				m_recvBuffer[i] = m_recvBuffer[i]+1;
+				m_recvBuffer[i] = (byte)(m_recvBuffer[i]+1);
 			}
 			serialPort1.Write(m_recvBuffer, 0, numBytes);
 			serialPort1.Flush();
